requirements.txt
openpyxl
selenium
allure-pytest==2.8.6
allure-python-commons==2.8.6
apipkg==1.5
atomicwrites==1.3.0
attrs==19.3.0
bcrypt==3.1.7
certifi==2019.9.11
cffi==1.13.1
chardet==3.0.4
colorama==0.4.1
cryptography==2.8
docopt==0.6.2
enum34==1.1.6
execnet==1.7.1
idna==2.8
more-itertools==7.2.0
namedlist==1.7
packaging==19.2
paramiko==2.6.0
pipreqs==0.4.9
pluggy==0.13.0
py==1.8.0
pycparser==2.19
PyMySQL==0.9.3
PyNaCl==1.3.0
pyparsing==2.4.2
pytest==5.2.1
pytest-forked==1.1.3
pytest-html==2.0.0
pytest-metadata==1.8.0
pytest-parallel==0.0.9
pytest-xdist==1.30.0
requests==2.22.0
six==1.12.0
thrift==0.11.0
urllib3==1.25.6
wcwidth==0.1.7
yarg==0.1.9
